Move more view actions from qaction to actions
